refactor: move page components
